requirements.txt
Flask
Flask-Sockets
gevent
gevent-websocket
greenlet
itsdangerous
Jinja2
MarkupSafe
redis
Werkzeug
django
gunicorn
django-heroku
python-socketio
aiohttp
cchardet
aiodns
aiodjango
requests